requirements.txt
numpy>=1.21.0
pandas>=1.3.0
tqdm>=4.61.2
pillow>=8.3.2
pydantic>=1.8.2
python-decouple>=3.6
matplotlib>=3.4.2
requests
scikit-learn>=0.24.2
psutil>=5.9.0
python-decouple>=3.6
